PHP version selector
Company Intranet web-based Tool used to identify which versions a
particular release has been built from, to the version being dispatched
to the client. There is a button which then links to an SSRS report
server that also displays all the job fixes (Summary Change Logs)
included for each version.